+++ packages/sdk/src/config/e2e.ts
@@ -29,6 +29,10 @@ export const e2eConfig: EthereumConfig = {
 		erc721: toAddress("0x74B4721599213fdC7eD0bb67D3B25a6fb3c74B61"),
 		erc1155: toAddress("0xF35BC450544C7d8c2559B4DAeb51E9617F39e7C6"),
 	},
+	cryptoPunks: {
+		marketContract: ZERO_ADDRESS,
+		wrapperContract: ZERO_ADDRESS,
+	},
 	weth: toAddress("0xc6f33b62a94939e52e1b074c4ac1a801b869fdb2"),
 	auction: toAddress("0xE091e5511815107072ed1C42d0d00f151D6197ed"),
 }