+++ packages/sdk/src/config/mumbai.ts
@@ -29,6 +29,10 @@ export const mumbaiConfig: EthereumConfig = {
 		erc721: toAddress("0xa85180a21786bA65b0778bE1cb5CBA5E5c6cD21d"),
 		erc1155: toAddress("0xAa9CD5834E0009902EeAA3FEfAc6A160e9A096b4"),
 	},
+	cryptoPunks: {
+		marketContract: ZERO_ADDRESS,
+		wrapperContract: ZERO_ADDRESS,
+	},
 	weth: toAddress("0xa6fa4fb5f76172d178d61b04b0ecd319c5d1c0aa"),
 	auction: ZERO_ADDRESS,
 }